Updated PubsubSchemaCapableIOProvider javadoc
